+++ core/php/AbeilleMainD.php
@@ -1810,7 +1810,7 @@
     $running = AbeilleTools::getRunningDaemons();
     $daemons = AbeilleTools::diffExpectedRunningDaemons($config, $running);
     logMessage('debug', 'Daemons status: '.json_encode($daemons));
-    if ($daemons["main"] > 1){
+    if ($daemons["maind"] > 1){
         logMessage('error', "Un démon 'main' est déja lancé ! ".json_encode($daemons));
         exit(3);
     }